third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvas-currentColor.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description("Test that the fill and stroke style attribute yield opaque black when set to 'currentcolor' on an OffscreenCanvasRenderingContext2D");

// Tests onstructor of OffscreenCanvas and length/width change
var aCanvas = new OffscreenCanvas(50, 50);
var ctx = aCanvas.getContext('2d');
ctx.fillStyle = 'red';
ctx.fillStyle = 'currentcolor'
shouldBe("ctx.fillStyle", "'#000000'");

ctx.strokeStyle = 'red';
ctx.strokeStyle = 'currentcolor'
shouldBe("ctx.strokeStyle", "'#000000'");

</script>
