third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvas-currentColor-expected.txt

Test that the fill and stroke style attribute yield opaque black when set to 'currentcolor' on an OffscreenCanvasRenderingContext2D

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.fillStyle is '#000000'
PASS ctx.strokeStyle is '#000000'
PASS successfullyParsed is true

TEST COMPLETE
